requirements.txt
av==14.2.0
dataclasses_json==0.6.7
eva_decord==0.6.1
Flask==3.1.0
flask_cors==5.0.1
huggingface_hub==0.29.2
hydra-core==1.3.2
imagesize==1.4.1
iopath==0.1.10
matplotlib==3.10.1
numpy==2.2.3
omegaconf==2.3.0
opencv_python==4.10.0.84
pandas==2.2.3
Pillow==11.1.0
pycocotools==2.0.8
pydicom==3.0.1
setuptools==59.6.0
# skimage==0.0
strawberry==3.0
submitit==1.5.2
tensordict==0.7.2
# torch==2.5.0+cu121
# torchvision==0.20.0+cu121
tqdm==4.66.5
